Move Updater.kt out of it's own package
